re scope bug fix
